Changed: better likelihood function for hard HD threshold, working good.

+++ src/lshf.cpp
@@ -1,4 +1,5 @@
 #include "lshf.hpp"
+#include <cstdint>
 
 LSHF::LSHF(uint8_t k, uint8_t h, uint32_t m)
   : k(k)
@@ -127,9 +128,12 @@ bool LSHF::check_compatible(lshf_sptr_t lshf)
     return true;
   if (!((lshf->m == m) && (lshf->h == h) && (lshf->k == k) && (lshf->npos_v == npos_v) &&
         (lshf->ppos_v == ppos_v))) {
-    std::cout << "m: " << m << "/" << lshf->m << std::endl;
-    std::cout << "h: " << h << "/" << lshf->h << std::endl;
-    std::cout << "k: " << k << "/" << lshf->k << std::endl;
+    std::cout << "m: " << static_cast<uint32_t>(m) << "/" << static_cast<uint32_t>(lshf->m)
+              << std::endl;
+    std::cout << "h: " << static_cast<uint32_t>(h) << "/" << static_cast<uint32_t>(lshf->h)
+              << std::endl;
+    std::cout << "k: " << static_cast<uint32_t>(k) << "/" << static_cast<uint32_t>(lshf->k)
+              << std::endl;
     std::cout << "ppos_v:";
     for (uint8_t i = 0; i < h; ++i) {
       std::cout << " " << static_cast<uint32_t>(ppos_v[i]) << "/"
